jecq/python/requirements.txt
faiss-cpu
numpy>=2
setuptools
packaging
swig